==> 1. Preprocessing & Cleaning.ipynb <==
# Analysis of Food Store Accessibility in Erie County, New York
## Erin Gregoire
## Fall 2024

### Dataset Collection and Justification

As an Erie County native, I have heard anecdotal relics of food deserts within the county, especially in the City of Buffalo. In this project, I have perfomed analysis on the available food stores, income and population within the municipalities located in Erie County to answer the following questions:
- Are there any areas that have a "food desert" in terms of lack of quality food stores?
- Is there a correlation between average income of a town and the quality of their food stores?
- Does living in the City of Buffalo differ with the Suburbs in terms of food store quality and accessibility?
- Can we predict whether a person is a resident of the City of Buffalo or a Suburb based on their income and local food stores?

### About the Data
The combination of these data sets will shows what the various socioeconomic classes has access to in terms of the type of groceries and accessibility to food stores. These datasets will also be useful in providing response and predictor variables when using machine learning techniques to predict whether a location is in the City of Buffalo or a Suburb. These datasets are cohesive due to the nature of the Zipcode column which will act as a unique indicator and link the data together seamlessly. 

#### Retail Food Stores
- Dataset by the New York State Department of Agriculture and Markets
- Provides Food Store Names, their Establishment Codes (Such as Store and Bakery), and their location
- Quality is controlled by a column called "Store Focus" which denotes the primary goal of the store
     - Grocery: Store is primarily for food products, includes food from all food groups (Produce, Dairy, Meat, and Dry/Canned Good)
     - Convenience: Gas stations, mini-marts, corner stores, prepared food stores
     - Multi-Purpose: Large name stores where groceries are only a small percent of products sold
     - Specialty: Store focuses on only one food group or product. Includes bakeries, meat/fish stores, artisan goods
     - Pharmacy: Pharmacies or drugs stores that may carry a small selection of grocery products
     - Other: Stores that carry prepared or snack foods that are not easily categorized above such as catering businesses.

#### Income and Tax Data
- Dataset by the US Department of Treasury and the IRS
- Provides income brackets and the number of returns conducted within each bracket
- Approximate population is accounted for by using the total number of tax returns for each zipcode

#### Anticipated Challenges
- Removing excess data from zipcodes that are not in Erie County
- Merging datasets smoothly

### Data Preprocessing and Cleaning

In this set, I will begin by loading the datasets one at a time and working with them until fully cleaned. Then, each dataset will be saved into their own file to be used again in Step 3.

import pandas as pd
import json
import sqlite3
import ast

erie_zipcodes = ['14001', '14004', '14006', '14010', '14025', '14026', '14027', '14030', '14031', '14032', '14033', '14034', '14035', '14038', '14043', '14047', '14051', '14052', '14055', '14057', '14059', '14061', '14068', '14069', '14072', '14075', '14080', '14085', '14086', '14091', '14102', '14110', '14111', '14127', '14134', '14139', '14140', '14141', '14150', '14151', '14169', '14170', '14201', '14202', '14203', '14204', '14205', '14206', '14207', '14208', '14209', '14210', '14211', '14212', '14213', '14214', '14215', '14216', '14217', '14128', '14219', '14220', '14221', '14222', '14223', '14224', '14225', '14226', '14227', '14228', '14231', '14233', '14240', '14241', '14260', '14261', '14263', '14264', '14265', '14267', '14269', '14270', '14272', '14273', '14276', '14280']

#### Retail Food Stores Data

stores_data = pd.read_csv('Retail_Food_Stores_20241119.csv')

stores_data.head()

#### Collect only food stores located in Erie County:

erie_stores = pd.DataFrame(stores_data[stores_data["County"] == "ERIE"])
erie_stores.head()

#### Remove unneccessary data

erie_stores = erie_stores.drop(columns=['County', 'License Number', 'Operation Type', 'Entity Name', 'Street Number', 'Street Name', 'Address Line 2', 'Address Line 3', 'State', 'Square Footage', 'Georeference'])

#### Rename columns and Check data type

erie_stores = erie_stores.rename(columns={'DBA Name': 'Food Store Name'})
erie_stores = erie_stores.rename(columns={'Zip Code': 'Zipcode'})

erie_stores.info()

erie_stores['Zipcode'] = erie_stores['Zipcode'].astype(str)

erie_stores.info()

#### Check for missing data and outliers

rows_with_missing_data = erie_stores[erie_stores.isna().any(axis=1)]
print(rows_with_missing_data)

establishment_types = erie_stores["Establishment Type"].unique()

for type in establishment_types:
    amount = len(erie_stores[erie_stores["Establishment Type"] == type])
    print (f' There are {amount} stores in Erie County with Establishment Type {type}')

There are no outliers specifically. However, there are some types of food stores that are unique. The majority of food stores fall into three main categories: AC (Store & Food Manufacturer), A (Store), and ABC (Store, Bakery, and Food Manufacturer). At this point, we do not need to remove or change any of the food store establishment types that are in unique categories. Many of the store types have overlap including the stores with unique types so they should still be relevant data.

#### Ensure that multi-valued attributes will be able to be handled in later steps

def split_types (types):
    result = list(types)
    return result

erie_stores['Establishment Type'] = erie_stores['Establishment Type'].apply(split_types)

len(erie_stores['Food Store Name'].unique())

erie_stores.head()

erie_stores.to_csv('erie_stores.csv', index=False)

#### Summary of Retail Food Store Data Cleaning Process
- Loaded CSV file
- Removed columns that belonged to counties other than Erie County
- Removed unnecessary columns that will not be useful to analysis
- Cleaned up column names, missing data, outliers, and data type discrepencies
- Saved file as new CSV

#### Income Tax Data

income_data = pd.read_csv('Tax_Zip_Code_Data.csv', usecols = (0, 1, 2, 3, 4, 5), skiprows = 3, skipfooter = 31, dtype = str, engine = 'python')

income_data.head(10)

#### Remove extra blank header rows and correct column names

income_data = income_data.drop(0).drop(1)

income_data.head(10)

income_data = income_data.rename(columns={'ZIP\ncode [1]': 'Zipcode'})

erie_incomes = income_data[income_data['Zipcode'].isin(erie_zipcodes)]

erie_incomes = erie_incomes.rename(columns={'Number of \nreturns [2]': 'Number of returns'})

erie_incomes.head(10)

#### Ensure Data Type Consistency

When the data was pulled into Python, pandas required specification for the data type of each column due to multiple data types in each column. They were all loaded in as a string for ease with the intention to correct them later in the data cleaning process. Below is how they were corrected.

erie_incomes.info()

erie_incomes[['Number of returns', 'Number of single returns', 'Number of joint returns', 'Number of head of household returns']] = erie_incomes[['Number of returns', 'Number of single returns', 'Number of joint returns', 'Number of head of household returns']].replace(',', '', regex=True)

erie_incomes.loc[erie_incomes['Number of returns'] == '** ', 'Number of returns'] = 0
erie_incomes.loc[erie_incomes['Number of single returns'] == '** ', 'Number of single returns'] = 0
erie_incomes.loc[erie_incomes['Number of joint returns'] == '** ', 'Number of joint returns'] = 0
erie_incomes.loc[erie_incomes['Number of head of household returns'] == '** ', 'Number of head of household returns'] = 0

erie_incomes.head(10)

erie_incomes[['Number of returns', 'Number of single returns', 'Number of joint returns', 'Number of head of household returns']] = erie_incomes[['Number of returns', 'Number of single returns', 'Number of joint returns', 'Number of head of household returns']].apply(pd.to_numeric)

erie_incomes.info()

erie_incomes = erie_incomes[erie_incomes["Zipcode"] != 0]

erie_incomes.head(20)

#### Feature Engineering

def calculate_percentage(group):
    total = group.iloc[0]["Number of returns"]
    group["Percentage (Number of returns)"] = (group["Number of returns"] / total) * 100
    return group

erie_incomes2 = erie_incomes.groupby("Zipcode", group_keys=False).apply(calculate_percentage, include_groups=False)

erie_incomes2["Zipcode"] = erie_incomes["Zipcode"]

cols = erie_incomes2.columns.tolist()

cols = cols[-1:] + cols[:-1]

erie_incomes = erie_incomes2[cols]

erie_incomes.info()

erie_incomes['Size of adjusted gross income'] = erie_incomes['Size of adjusted gross income'].fillna('Total')

erie_incomes.head()

erie_incomes.to_csv('erie_incomes.csv', index=False)

#### Summary of Income Tax Return Data Cleaning Process
Loaded CSV file
Removed rows that had blanks/missing data
Ensured data type consistency by converting rows to correct data types
Removed unnecessary columns that will not be useful to analysis
Removed rows that belonged to zipcodes not in Erie County
Feature engineered columns to calculate the percentage of tax returns for each income bracket
Saved to new CSV file

#### Establishment Type data from the Retail Food Store dataset

establishment_types = {
    "A" : 'Store',
    'B' : "Bakery",
    'C' : "Food Manufacturer",
    'D' : 'Food Warehouse',
    'H' : 'Wholesale Manufactuer',
    'K' : 'Vehicle',
    'W' : 'Farm Winery',
    'Y' : 'Slaughterhouse'}

import json

with open('establishments.json', 'w', encoding='utf-8') as file:
        json.dump(establishment_types, file, sort_keys=True, indent=4, separators=(',', ': '), ensure_ascii=False)

with open('establishments.json', 'r', encoding='utf-8') as file:
    testing = json.load(file)

#### Summary of Establishment Type Data Cleaning Process
- In order to understand the meaning of the establishment types, created a dictionary to hold the code and the meaning as key, value pairs
- Saved dictionary as a JSON file

### Summary of Data Cleaning Process:
- Food Store, Income Tax, and Establishment Type data are now stored in their own files
- Data only includes data for zipcodes in Erie County
- Zipcodes are saved as string data types due to their purpose at categorizing data, rather than being used as an integer with functions
- Impact of Cleaning:
    - Drastically reduced the size of the data we are working with, both by removing rows with zipcodes in counties other than Erie and also removing columns that provide irrelevant data.
    - Data is provided in columns and tables that are easy to understand and also very accessible
