==> 2. Database Construction.ipynb <==
# Analysis of Food Store Accessibility in Erie County, New York
## Erin Gregoire
## Fall 2024

import pandas as pd
import sqlite3
import ast

### SQL Database Design

In this step, I have created a database called "Erie County DB" to hold all of the data for this project. This step also details the process of normalization

conn = sqlite3.connect('eriecountydb')
cur = conn.cursor()

store_data = pd.read_csv('erie_stores.csv')
income_data = pd.read_csv('erie_incomes.csv')

One challenge that was unanticipated was that there is the same zipcode but called multiple different town names or if a town name was mispelled in the original data. Here I have removed any duplicate names that have appeared. 

unique_pairs = []
zips = store_data['Zipcode']
cities = store_data['City']
pairs = zip(zips, cities)

for pair in pairs:
    if pair not in unique_pairs:
        unique_pairs.append(pair)

unique_pairs.remove((14043, 'LANCASTER'))
unique_pairs.remove((14085, 'LAKEVIEW'))
unique_pairs.remove((14150, 'TONWANDA'))
unique_pairs.remove((14206, 'CHEEKTOWAGA'))
unique_pairs.remove((14211, 'CHEEKTOWAGA'))
unique_pairs.remove((14215, 'CHEEKTOWAGA'))
unique_pairs.remove((14217, 'BUFFALO'))
unique_pairs.remove((14218, 'BUFFALO'))
unique_pairs.remove((14224, 'W SENECA'))
unique_pairs.remove((14219, 'BUFFALO'))
unique_pairs.remove((14221, 'BUFFALO'))
unique_pairs.remove((14223, 'BUFFALO'))
unique_pairs.remove((14225, 'BUFFALO'))
unique_pairs.remove((14226, 'BUFFALO'))
unique_pairs.remove((14227, 'BUFFALO'))
unique_pairs.remove((14228, 'BUFFALO'))
unique_pairs.remove((14221, 'AMHERST'))
unique_pairs.remove((14221, 'CLARENCE'))
unique_pairs.remove((14224, 'CHEEKTOWAGA'))
unique_pairs.remove((14224, 'BUFFALO'))

z = []
c = []
for p in unique_pairs:
    z.append(p[0])
    c.append(p[1])

#### Location Table
- Primary Key: Zipcode
- 1NF: Removed duplicated of town names so that data is atomic
- 2NF/3NF: N/A

cur.execute("DROP TABLE IF EXISTS Location")

cur.execute('''
        CREATE TABLE Location (
        Zipcode TEXT PRIMARY KEY,
        City TEXT)
        ''')

cur.executemany("""INSERT INTO Location (Zipcode, City) VALUES (?, ?)""", zip(z, c))

pd.read_sql_query('''SELECT * FROM Location''', conn)

#### Establishment Type Table
- Primary Key: Establishment Code (Est_Code)
- 3NF: Establishment Type was given its own table to store the Establishment Type Data. Without its own table, there would be transitive dependency of the Establishment Type on the Establishment Code if remained connected to the Food Store table.
- 1NF/2NF: N/A

with open('establishments.json', 'r', encoding='utf-8') as file:
    est_data = json.load(file)

cur.execute("DROP TABLE IF EXISTS Establishment_Type")

cur.execute('''
        CREATE TABLE Establishment_Type (
        Est_Code TEXT PRIMARY KEY,
        Est_Type TEXT
        )
        ''')

cur.executemany("""INSERT INTO Establishment_Type (Est_Code, Est_Type) VALUES (?, ?)""", zip(list(est_data.keys()), list(est_data.values())))

pd.read_sql_query('''SELECT * FROM Establishment_Type''', conn)

#### Food Store Table
- Primary Key: Store_ID
- Foreign Key: Zipcode, Establishment Code
- 1NF: Created Store_ID due to multiple stores having the same name. Establishment codes are split into multiple columns to remove multi-valued data
- 3NF: Establishment Type removed which was dependent on Establishment Code.
- 2NF: N/A

cur.execute("DROP TABLE IF EXISTS Food_Store")

cur.execute('''
        CREATE TABLE Food_Store (
        Store_ID INTEGER PRIMARY KEY,
        Store_Name TEXT,
        Zipcode TEXT,
        Store_Focus,
        Est_Code1 TEXT,
        Est_Code2 TEXT,
        Est_Code3 TEXT,
        Est_Code4 TEXT,
        UNIQUE (Store_ID),
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode),
        FOREIGN KEY (Est_Code1) REFERENCES Establishment_Type (Est_Code),
        FOREIGN KEY (Est_Code2) REFERENCES Establishment_Type (Est_Code),
        FOREIGN KEY (Est_Code3) REFERENCES Establishment_Type (Est_Code),
        FOREIGN KEY (Est_Code4) REFERENCES Establishment_Type (Est_Code)
        )
        ''')

info = list(zip(store_data['Food Store Name'], store_data['Zipcode'], store_data['Establishment Type']))

info2 = []
for tup in info:
    lst = ast.literal_eval(tup[2])
    info2.append(lst)

est_code1 = []
est_code2 = []
est_code3 = []
est_code4 = []
for code in info2:
    if len(code) == 1:
        est_code1.append(code[0])
        est_code2.append(None)
        est_code3.append(None)
        est_code4.append(None)
    elif len(code) == 2:
        est_code1.append(code[0])
        est_code2.append(code[1])
        est_code3.append(None)
        est_code4.append(None)
    elif len(code) == 3:
        est_code1.append(code[0])
        est_code2.append(code[1])
        est_code3.append(code[2])
        est_code4.append(None)
    elif len(code) == 4:
        est_code1.append(code[0])
        est_code2.append(code[1])
        est_code3.append(code[2])
        est_code4.append(code[3])

with open('storefocus1.txt', 'r') as f:
    focus_dat = f.read()

focus = focus_dat.strip()
focus = focus.replace('ï»¿Convenience', 'Convenience')
focus = focus.split('\n')

cur.executemany("""INSERT INTO Food_Store (Store_Name, Zipcode, Store_Focus, Est_Code1, Est_Code2, Est_Code3, Est_Code4) VALUES (?, ?, ?, ?, ?, ?, ?)""", zip(store_data['Food Store Name'], store_data['Zipcode'], focus, est_code1, est_code2, est_code3, est_code4))

pd.read_sql_query('''SELECT * FROM Food_Store''', conn)

#### Income Tables:
- Primary Key: Zipcode
- Foreign Key: Zipcode
- 2NF: Removed partial dependencies by creating a seperate table for each income bracket
- 1NF/3NF: N/A

income_data.columns

data = list(zip(income_data['Zipcode'], income_data['Size of adjusted gross income'], income_data['Number of returns'], income_data['Percentage (Number of returns)']))

data2 = []
for t in data:
    if t[0] in list(store_data['Zipcode']):
        data2.append(t)

total = []
under25 = []
_25to50 = []
_50to75 = []
_75to100 = []
_100to200 = []
above200 = []

for i in data2:
    if i[1] == 'Total':
        total.append(i)
    elif i[1] == '$1 under $25,000':
        under25.append(i)
    elif i[1] == '$25,000 under $50,000':
        _25to50.append(i)
    elif i[1] == '$50,000 under $75,000':
        _50to75.append(i)
    elif i[1] == '$75,000 under $100,000':
        _75to100.append(i)
    elif i[1] == '$100,000 under $200,000':
        _100to200.append(i)
    elif i[1] == '$200,000 or more':
        above200.append(i)

zip_total = []
num_total = []
for tot in total:
    zip_total.append(tot[0])
    num_total.append(tot[2])

##### Income Total Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

cur.execute("DROP TABLE IF EXISTS Income_Total")

cur.execute('''
        CREATE TABLE Income_Total (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_Total (Zipcode, No_of_Returns) VALUES (?, ?)""", zip(zip_total, num_total))

pd.read_sql_query('''SELECT * FROM Income_Total''', conn)

##### Income Under 25k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_under25 = []
num_under25 = []
perc_under25 = []
for u in under25:
    zip_under25.append(u[0])
    num_under25.append(u[2])
    perc_under25.append(round(u[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_Under_25k")

cur.execute('''
        CREATE TABLE Income_Under_25k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_Under_25k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_under25, num_under25, perc_under25))

pd.read_sql_query('''SELECT * FROM Income_Under_25k''', conn)

##### Income for 25k to 50k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_25 = []
num_25 = []
perc_25 = []
for v in _25to50:
    zip_25.append(v[0])
    num_25.append(v[2])
    perc_25.append(round(v[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_25k_to_50k")

cur.execute('''
        CREATE TABLE Income_25k_to_50k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_25k_to_50k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_25, num_25, perc_25))

pd.read_sql_query('''SELECT * FROM Income_25k_to_50k''', conn)

##### Income for 50k to 75k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_50 = []
num_50 = []
perc_50 = []
for w in _50to75:
    zip_50.append(w[0])
    num_50.append(w[2])
    perc_50.append(round(w[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_50k_to_75k")

cur.execute('''
        CREATE TABLE Income_50k_to_75k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_50k_to_75k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_50, num_50, perc_50))

pd.read_sql_query('''SELECT * FROM Income_50k_to_75k''', conn)

##### Income 75k to 100k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_75 = []
num_75 = []
perc_75 = []
for nn in _75to100:
    zip_75.append(nn[0])
    num_75.append(nn[2])
    perc_75.append(round(nn[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_75k_to_100k")

cur.execute('''
        CREATE TABLE Income_75k_to_100k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_75k_to_100k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_75, num_75, perc_75))

pd.read_sql_query('''SELECT * FROM Income_75k_to_100k''', conn)

##### Income for 100k to 200k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_100 = []
num_100 = []
perc_100 = []
for mm in _100to200:
    zip_100.append(mm[0])
    num_100.append(mm[2])
    perc_100.append(round(mm[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_100k_to_200k")

cur.execute('''
        CREATE TABLE Income_100k_to_200k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_100k_to_200k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_100, num_100, perc_100))

pd.read_sql_query('''SELECT * FROM Income_100k_to_200k''', conn)

##### Income Above 200k Table
- Primary Key: Zipcode
- Foreign Key: Zipcode

zip_200 = []
num_200 = []
perc_200 = []
for zz in above200:
    zip_200.append(zz[0])
    num_200.append(zz[2])
    perc_200.append(round(zz[3], 2))

cur.execute("DROP TABLE IF EXISTS Income_Above_200k")

cur.execute('''
        CREATE TABLE Income_Above_200k (
        Zipcode TEXT PRIMARY KEY,
        No_of_Returns INTEGER,
        Percentage REAL,
        FOREIGN KEY (Zipcode) REFERENCES Location (Zipcode)
        )
        ''')

cur.executemany("""INSERT INTO Income_Above_200k (Zipcode, No_of_Returns, Percentage) VALUES (?, ?, ?)""", zip(zip_200, num_200, perc_200))

pd.read_sql_query('''SELECT * FROM Income_Above_200k''', conn)

conn.commit()
conn.close()

### Summary of SQL Database Design
- All tables have been normalized to 3NF
- Tables within the Erie County Database:
    - Location
    - Food Store
    - Establishment Type
    - Income Total
    - Income Under 25k
    - Income 25k to 50k
    - Income 50k to 75k
    - Income 75k to 100k
    - Income 100k to 200k
    - Income Above 200k
